Add abstract class for functional tests
